+++ packages/api/package.json
@@ -2,7 +2,7 @@
     "name": "@impactmarket/api",
     "main": "./dist/app.js",
     "version": "1.0.0",
-    "license": "MIT",
+    "license": "Apache-2.0",
     "engines": {
         "node": ">=16.0.0 <17.0.0"
     },
@@ -21,15 +21,16 @@
         "@types/jsonwebtoken": "8.5.6",
         "@types/mocha": "9.0.0",
         "@types/multer": "^1.4.7",
+        "@types/node": "16.18.10",
         "@types/rate-limit-redis": "1.7.4",
         "mocha": "9.1.3",
-        "nodemon": "2.0.15"
+        "nodemon": "2.0.15",
+        "openapi-types": "7.2.3"
     },
     "dependencies": {
         "@impactmarket/core": "1.0.0",
-        "@sentry/node": "6.17.8",
-        "@sentry/tracing": "6.17.8",
-        "@slack/events-api": "3.0.1",
+        "@sentry/node": "7.27.0",
+        "@sentry/tracing": "7.27.0",
         "bignumber.js": "9.0.0",
         "body-parser": "1.19.0",
         "celebrate": "15.0.0",
@@ -38,7 +39,7 @@
         "ethers": "5.5.1",
         "express": "4.17.3",
         "express-rate-limit": "5.5.1",
-        "helmet": "4.6.0",
+        "helmet": "6.0.1",
         "json2csv": "5.0.7",
         "jsonwebtoken": "8.5.1",
         "module-alias": "2.2.2",
